Fixed multiple issues related to ** query handling.
dpath had some unexpected behavior. When the query "**" was passed over
a range, the items in that range would be returned, along with the
dictionaries and lists. This is only what I think can be called Russian
Doll style.

This was not wanted, so a filter was put in place too make sure only the
items are returned.

Also, when this occured it caused an error in flatten table, as it
failed to handle this case. A case handler for this was added (with a
print message as this would be unexpected behavior).

Also, a catch for gspread errors was added, if the attempt to add a
list, caused the sheet to object because of exceeding the 2 million cell
limit.